% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ang2rot.R
\name{ang2rot}
\alias{ang2rot}
\title{Transforms from angles in the range [-pi,pi] to angles of rotation, so that when circling around an object, the angle increases beyond this range.}
\usage{
ang2rot(ang, tol = 5)
}
\arguments{
\item{ang}{is a vector af angles as returned from atan2.}

\item{tol}{is the angle difference defined as representing change from -pi to pi, or pi to -pi.}
}
\value{

}
\description{
Transforms from angles in the range [-pi,pi] to angles of rotation, so that when circling around an object, the angle increases beyond this range.
}
\examples{
\dontrun{}

}
