% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arr.ind2ind.R
\name{arr.ind2ind}
\alias{arr.ind2ind}
\title{Transforms array indexes to vector indexes. For indexes out of bounds NA is returned. The function is transformed from the function which().}
\usage{
arr.ind2ind(arr.ind, shape, allow.outside = FALSE)
}
\arguments{
\item{arr.ind}{is either a list of subscripts, as typed into [], or an array indexes, to be transformed to vector indexes.}

\item{shape}{is the dimension of the array.}
}
\value{

}
\description{
Transforms array indexes to vector indexes. For indexes out of bounds NA is returned. The function is transformed from the function which().
}
\examples{
\dontrun{}

}
