% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchWarnings.R
\name{catchWarnings}
\alias{catchWarnings}
\title{It returns a 2-element list, the first being the value of the expression given to it and the second being a list of all the warnings. catchWarnings() is based on suppressWarnings() and used similarly. Created by Bill Dunlap , TIBCO Software Inc - Spotfire Division , wdunlap tibco.com.}
\usage{
catchWarnings(expr)
}
\arguments{
\item{expr}{is an R-expression to be evaluated.}
}
\value{

}
\description{
It returns a 2-element list, the first being the value of the expression given to it and the second being a list of all the warnings. catchWarnings() is based on suppressWarnings() and used similarly. Created by Bill Dunlap , TIBCO Software Inc - Spotfire Division , wdunlap tibco.com.
}
\examples{
\dontrun{}

}
