% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.TSD.R
\name{combine.TSD}
\alias{combine.TSD}
\title{Merges either (1) TSD files given by 'x', or (2) four character named TSD variables in the list 'x', as returned when reading mutiple TSD files. List elements of identical names are collapsed.}
\usage{
combine.TSD(
  x,
  dir = NULL,
  indt = FALSE,
  reserve = TRUE,
  recursive = FALSE,
  test.TSD = FALSE,
  filesize = 3e+08,
  chunksize = 3e+08,
  clear_along = FALSE,
  drop.out = FALSE,
  adds = NULL,
  msg = TRUE,
  linked = list(),
  skipLast = FALSE,
  cores = 1,
  keep.lists = FALSE,
  pad = TRUE,
  pbar = TRUE,
  ...
)
}
\arguments{
\item{x}{A list of TSD-data, or a directory or vector of TSD-files to be merged, in which case a new directory is created with the merged data.}

\item{dir}{The directory in which to put the merged file(s).}

\item{indt}{Used then merging TSD-data in the list 'x', causing values for duplicated time steps to be discarded (the variable 'indt' must be present, one for each file/group of data).}

\item{reserve}{Logical: If FALSE time steps should not be reserved, requiring that the lengths and dimensions of the variables in the files to merge match exactly.}

\item{recursive}{Used when listing the files to be merged, and if set to TRUE files in subfolders will be included as well.}

\item{test.TSD}{Logical: If TRUE discard files that are not TSD files (could be time consuming).}

\item{filesize}{The maximum size of the merged files.}

\item{chunksize}{The maximum size of the chunks of file read at the time.}

\item{clear_along}{Logical: If TRUE clear files that have been merged imediately after merging.}

\item{drop.out}{Logical: If TRUE drop dimensions of the output.}

\item{adds}{An optional list of variables added before merging.}

\item{msg}{Logical: If TRUE print progress and messages to the console.}

\item{linked}{A list of file paths of the same length as 'x' (or the file list of 'x'), holding files that should be merged using the same file grouping as used for 'x'.}

\item{skipLast}{Logical: If TRUE discard the last file in the merging, and simply copy it to the merged-directory.}

\item{cores}{The number of cores to run parallel mering on.}

\item{keep.lists}{Logical: If TRUE, keep the output as lists.}

\item{pad}{See \code{\link{mergeListKeepDimensions}}.}

\item{pbar}{Logical: If FALSE, the progress bar is suppressed.}

\item{...}{Possible inputs to other functions.}
}
\description{
Merges either (1) TSD files given by 'x', or (2) four character named TSD variables in the list 'x', as returned when reading mutiple TSD files. List elements of identical names are collapsed.
}
\examples{
\dontrun{}

}
