% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_all.R
\name{dim_all}
\alias{dim_all}
\title{Returns the dimension of the object 'x'. If 'x' is a list, a list of dimensions is returned. If 'x' is a vector length(x) is returned. If 'x' is NULL, 'null.out' is returned.}
\usage{
dim_all(x, null.out = NULL, old = TRUE)
}
\arguments{
\item{x}{The input object.}

\item{null.out}{The value to return from NULL.}

\item{old}{Deprecated. Logical: If TRUE use the old function.}
}
\description{
Returns the dimension of the object 'x'. If 'x' is a list, a list of dimensions is returned. If 'x' is a vector length(x) is returned. If 'x' is NULL, 'null.out' is returned.
}
\examples{
\dontrun{}

}
