% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.callUnique.R
\name{do.callUnique}
\alias{do.callUnique}
\title{Execute a Function Call, allowing for further arguments given as ...}
\usage{
do.callUnique(what, args, ...)
}
\arguments{
\item{what}{either a function or a non-empty character string naming the function to be called.}

\item{args}{a list of arguments to the function call. The names attribute of args gives the argument names.}

\item{...}{futher agruments to be merged with 'args'. These arguments have presidence over those in 'args'.}
}
\description{
Execute a Function Call, allowing for further arguments given as ...
}
