% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/even.R
\name{even}
\alias{even}
\alias{odd}
\title{Returning TRUE at all even or odd values.}
\usage{
even(x)

odd(x)
}
\arguments{
\item{x}{is the R-object to be checked for even or odd values.}
}
\description{
Returning TRUE at all even or odd values.
}
\examples{
\dontrun{}

}
