% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIndSubset.R
\name{extractIndSubset}
\alias{extractIndSubset}
\title{Extracts a subset of 'x' defined by the index list or vector 'ind' and/or the logical vector 'subset'.}
\usage{
extractIndSubset(
  x,
  ind = list(),
  subset = NULL,
  drop = TRUE,
  ind.out = FALSE,
  insert.NA = FALSE,
  only.match = FALSE,
  return.all = TRUE,
  pad = c("end", "start")
)
}
\arguments{
\item{x}{is the list or array to be extracted from.}

\item{ind}{is a list of indexes, as typed into the [] of an array, where 0 and NULL denotes all indexes.}

\item{subset}{is a numeric or logical vector/expression indicating elements or rows to keep. Missing values are taken as false, and subset=0 or subset=NULL indicates no subsetting. Can be given as a list with subsetting indices for each time step.}

\item{drop}{is TRUE if dimensions of only one level is to be removed from the output.}

\item{ind.out}{is TRUE if the one dimensinal vector indexes used in the extraction is to be returned.}

\item{insert.NA}{is TRUE if the discarded data are to be kept as NA.}

\item{only.match}{is TRUE if only the arrays of length equal to the length of 'subset' are to be subsetted using 'subset'.}

\item{return.all}{is FALSE to return empty data and indices when 'ind' and 'subset' are enpty.}

\item{pad}{determines at which end of the dimensions to pad with zeros if the length of 'ind' is shorter than the number of dimensions. Using pad="start" applies 'ind' on the last dimensions.}
}
\description{
Extracts a subset of 'x' defined by the index list or vector 'ind' and/or the logical vector 'subset'.
}
\examples{
\dontrun{}

}
