% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2character.R
\name{function2character}
\alias{function2character}
\title{Converts a function to a string starting with " <- function", which can be sourced by R if pasted a name for the function prior to the string.}
\usage{
function2character(x, fname = "")
}
\arguments{
\item{x}{is the function.}

\item{fname}{is the optional function name.}
}
\description{
Converts a function to a string starting with " <- function", which can be sourced by R if pasted a name for the function prior to the string.
}
\examples{
\dontrun{}

}
