% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind2arr.ind.R
\name{ind2arr.ind}
\alias{ind2arr.ind}
\title{Transforms vector indexes to array indexes. The function is cloned from the function which().}
\usage{
ind2arr.ind(ind, shape = NULL, perm = NULL, ...)
}
\arguments{
\item{ind}{The index(es) to be transformed to array indexes.}

\item{shape}{The dimension of the array.}

\item{perm}{The permutation i. e. the order of the dimensions along which the indexes given in 'ind' are organized. If 'shape' is 2:4 and 'perm' is c(3,1,2), the indexes in 'ind' are first organized along the third dimension, then along the first, and finally along the second dimension of a matrix of dimensions 2:4.}

\item{...}{Used for robustness.}
}
\value{

}
\description{
Transforms vector indexes to array indexes. The function is cloned from the function which().
}
\examples{
\dontrun{}

}
