% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.charTSD_element.R
\name{is.charTSD_element}
\alias{is.charTSD_element}
\title{Simple function for determining whether an element of the data 'y' is of character type.}
\usage{
is.charTSD_element(y)
}
\arguments{
\item{y}{is an R object.}
}
\value{

}
\description{
Simple function for determining whether an element of the data 'y' is of character type.
}
\examples{
\dontrun{}

}
