% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.complexTSD_element.R
\name{is.complexTSD_element}
\alias{is.complexTSD_element}
\title{Simple function for determining whether an element of the data 'y' is of complex type.}
\usage{
is.complexTSD_element(y)
}
\arguments{
\item{y}{is an R object.}
}
\value{

}
\description{
Simple function for determining whether an element of the data 'y' is of complex type.
}
\examples{
\dontrun{}

}
