% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.doubleTSD_element.R
\name{is.doubleTSD_element}
\alias{is.doubleTSD_element}
\title{Simple function for determining whether an element of the data 'y' is too large to be represented as float with presicion for whole numbers.}
\usage{
is.doubleTSD_element(y)
}
\arguments{
\item{y}{is an R object.}
}
\description{
Simple function for determining whether an element of the data 'y' is too large to be represented as float with presicion for whole numbers.
}
\examples{
\dontrun{}

}
