% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.integerTSD_element.R
\name{is.integerTSD_element}
\alias{is.integerTSD_element}
\title{Simple function for determining whether an element of the data 'y' is of integer type.}
\usage{
is.integerTSD_element(y)
}
\arguments{
\item{y}{is an R object.}
}
\description{
Simple function for determining whether an element of the data 'y' is of integer type.
}
\examples{
\dontrun{}

}
