% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.TSD.R
\name{is.TSD}
\alias{is.TSD}
\title{Tests whether the file(s) given by 'x' has the Time Step Data format (TSD) by reading the first four bytes as character and comparing to the TSD file header "\%TSD".}
\usage{
is.TSD(x, recursive = TRUE, accept.dir = TRUE)
}
\arguments{
\item{x}{is a directory or a vector of file paths.}

\item{recursive}{is used in list.files() when 'x' is a directory.}

\item{accept.dir}{is TRUE to permit giving 'x' as a directory, which is only used for a special reason in read.TSD().}
}
\value{

}
\description{
Tests whether the file(s) given by 'x' has the Time Step Data format (TSD) by reading the first four bytes as character and comparing to the TSD file header "\%TSD".
}
\examples{
\dontrun{}

}
