% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labl.TSD.R
\name{labl.TSD}
\alias{labl.TSD}
\title{Returns the variable names required in echoIBM() and related functions, given as four character strings of names as stored in the TSD format. The aim of this function is to avoid errors due to misspelling.}
\usage{
labl.TSD(var = NULL, adds = NULL, list.out = TRUE)
}
\arguments{
\item{var}{is the variable classes for which variable names are requested. Implemented variable classes are:}

\item{adds}{is a vector of additional variables to include.}
}
\value{

}
\description{
Returns the variable names required in echoIBM() and related functions, given as four character strings of names as stored in the TSD format. The aim of this function is to avoid errors due to misspelling.
}
\examples{
\dontrun{}

}
