% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len_all.R
\name{len_all}
\alias{len_all}
\title{Returns the lengths of the object 'x'.}
\usage{
len_all(x)
}
\arguments{
\item{x}{is the input object.}
}
\value{

}
\description{
Returns the lengths of the object 'x'.
}
\examples{
\dontrun{}

}
