% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOfEqual2array.R
\name{listOfEqual2array}
\alias{listOfEqual2array}
\title{(Internal) Convert a list of equal lengths to an array.}
\usage{
listOfEqual2array(x)
}
\arguments{
\item{x}{A list.}
}
\value{

}
\description{
(Internal) Convert a list of equal lengths to an array.
}
\examples{
\dontrun{}

}
