% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2dhm.R
\name{m2dhm}
\alias{m2dhm}
\title{Transforms minutes to an array of days, hours and minutes.}
\usage{
m2dhm(m, digits = 3, clock.out = FALSE, names = c("d", "h", "m"),
  strip = FALSE, drop = FALSE)
}
\arguments{
\item{m}{is the input minutes.}

\item{digits}{is the number of digits of the "seconds" part of the output.}

\item{clock.out}{is TRUE if the time excluding the days are to be returned as a string vector of elements HH:MM:SS.FFF (two digits).}
}
\value{

}
\description{
Transforms minutes to an array of days, hours and minutes.
}
\examples{
\dontrun{}

}
