% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_TSD.R
\name{merge_TSD}
\alias{merge_TSD}
\title{Merges either (1) TSD files given by 'x', or (2) four character named TSD variables in the list 'x', as returned when reading mutiple TSD files. List elements of identical names are collapsed.}
\usage{
merge_TSD(x, dir = NULL, indt = FALSE, reserve = TRUE,
  recursive = FALSE, test.TSD = FALSE, filesize = 3e+08,
  chunksize = 3e+08, clear_along = FALSE, drop.out = FALSE, adds = NULL,
  msg = TRUE, linked = list(), skipLast = FALSE, cores = 1,
  keep.lists = FALSE, pad = TRUE, ...)
}
\arguments{
\item{x}{is a list of TSD-data, or a directory or vector of TSD-files to be merged, in which case a new directory is created with the merged data.}

\item{dir}{is the directory in which to put the merged file(s).}

\item{indt}{is used then merging TSD-data in the list 'x', causing values for duplicated time steps to be discarded (the variable 'indt' must be present, one for each file/group of data).}

\item{reserve}{is FALSE if time steps should not be reserved, requiring that the lengths and dimensions of the variables in the files to merge match exactly.}

\item{recursive}{is used when listing the files to be merged, and if set to TRUE files in subfolders will be included as well.}

\item{test.TSD}{is TRUE to discard files that are not TSD files (could be time consuming).}

\item{filesize}{is the maximum size of the merged files.}

\item{chunksize}{is the maximum size of the chunks of file read at the time.}

\item{clear_along}{is TRUE to clear files that have been merged imediately after merging.}

\item{drop.out}{is TRUE to clear files that have been merged imediately after merging.}

\item{adds}{is TRUE to clear files that have been merged imediately after merging.}

\item{linked}{is a list of file paths of the same length as 'x' (or the file list of 'x'), holding files that should be merged using the same file grouping as used for 'x'.}

\item{skipLast}{is TRUE to discard the last file in the merging, and simply copy is to the merged-directory.}

\item{...}{are possible inputs to other functions.}
}
\value{

}
\description{
Merges either (1) TSD files given by 'x', or (2) four character named TSD variables in the list 'x', as returned when reading mutiple TSD files. List elements of identical names are collapsed.
}
\examples{
\dontrun{}

}
