% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeListKeepDimensions.R
\name{mergeListKeepDimensions}
\alias{mergeListKeepDimensions}
\title{Funciton used for merging data stored in a list, where each list element contains data from one file}
\usage{
mergeListKeepDimensions(x, pad = TRUE, add1 = FALSE, split = TRUE)
}
\arguments{
\item{x}{is a list containing some data that should be merged, and that preferably has equal dimension up to the last one (which might be missing). If not equal in dimension, 'pad' can be set to TRUE to pad with missing values.}

\item{pad}{is TRUE to pad with missing values to make the list elements fit in the dimension up to the last dimension.}

\item{add1}{Logical: if TRUE, add one empty dimension to all elements of the data.}

\item{split}{Logical: if TRUE, split each list element to have last dimenstion equal to 1.}
}
\value{

}
\description{
Funciton used for merging data stored in a list, where each list element contains data from one file
}
\examples{
\dontrun{}

}
