% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTextFiles.R
\name{mergeTextFiles}
\alias{mergeTextFiles}
\title{Merges text files given a maximum size}
\usage{
mergeTextFiles(files, con = NULL, maxsize = 1e+06)
}
\arguments{
\item{files}{A vector of file paths.}

\item{con}{The file path of the merged file, appended by integers before the dot if more than one file is written.}

\item{maxsize}{Maximum size of the merged files.}
}
\description{
Merges text files given a maximum size
}
