% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{mod}
\alias{mod}
\title{(Internal) Returns the modulus of x by y, but keeping y when x<percent><percent>y returns 0.}
\usage{
mod(x, y)
}
\arguments{
\item{x, y}{The input vectors.}
}
\value{

}
\description{
(Internal) Returns the modulus of x by y, but keeping y when x<percent><percent>y returns 0.
}
\examples{
\dontrun{}

}
