% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numt.TSD.R
\name{numt.TSD}
\alias{numt.TSD}
\title{Get the number of time steps.}
\usage{
numt.TSD(x)
}
\arguments{
\item{x}{is a list containing the required variables 'dirx' (vector of x-beam-directions), 'diry' (vector of y-beam-directions), 'dirz' (vector of z-beam-directions), and 'esnm' (name of the echosounder/sonar).}
}
\description{
Get the number of time steps.
}
