% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathparts.R
\name{pathparts}
\alias{pathparts}
\title{Splits file paths into parts. If only one file path is given, a vector is returned. If multiple file paths are given in a vector, a list is returned.}
\usage{
pathparts(x)
}
\arguments{
\item{x}{is a vector of file paths.}
}
\value{

}
\description{
Splits file paths into parts. If only one file path is given, a vector is returned. If multiple file paths are given in a vector, a list is returned.
}
\examples{
\dontrun{}

}
