% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyIntegers.R
\name{prettyIntegers}
\alias{prettyIntegers}
\title{Compressed formating of an integer vector.}
\usage{
prettyIntegers(x, force = FALSE, sep = "...", collapse = ", ")
}
\arguments{
\item{x}{is an integer vector.}

\item{force}{is TRUE if the new string should be returned regardless of whether it has more characters than the original.}
}
\value{

}
\description{
Compressed formating of an integer vector.
}
\examples{
\dontrun{}

}
