% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.TSD_description_table.R
\name{read.TSD_description_table}
\alias{read.TSD_description_table}
\title{Reads the TSD description table 'x'.}
\usage{
read.TSD_description_table(x)
}
\arguments{
\item{x}{is the file containing a table of two columns named "Label" and "Description", specifying in the first column the TSD names, which are UTF-8 four character names, and in the second column the descriptions of the variables in the first column.}
}
\description{
Reads the TSD description table 'x'.
}
\examples{
\dontrun{}

}
