% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.TSD_funstring2fun.R
\name{read.TSD_funstring2fun}
\alias{read.TSD_funstring2fun}
\title{Simple function for extracting the function saved as a string using function2character().}
\usage{
read.TSD_funstring2fun(x)
}
\arguments{
\item{x}{is a function string.}
}
\value{

}
\description{
Simple function for extracting the function saved as a string using function2character().
}
\examples{
\dontrun{}

}
