% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.TSD_isnvar.R
\name{read.TSD_isnvar}
\alias{read.TSD_isnvar}
\title{Function identifying whether a sequence of bytes of length 4*n at the current position is a float or not, and storing the item read.}
\usage{
read.TSD_isnvar(con, n = 1L, endian = .Platform$endian, max_var = 1e+10,
  fixed_endian = FALSE)
}
\arguments{
\item{con}{is a connection to a binary file, as returned by file(x,"rb").}

\item{n}{is the number of elements to read.}

\item{endian}{is the endianness of the file, changed if the float is not read properly.}

\item{max_var}{is the maximum value accepted as float.}

\item{fixed_endian}{is TRUE if the endian should not be swaped, and readable set to FALSE for wrong endian.}
}
\value{

}
\description{
Function identifying whether a sequence of bytes of length 4*n at the current position is a float or not, and storing the item read.
}
\examples{
\dontrun{}

}
