% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.TSD_readAllTimeStepsOfOneVariable.R
\name{read.TSD_readAllTimeStepsOfOneVariable}
\alias{read.TSD_readAllTimeStepsOfOneVariable}
\title{(Internal) Read all time steps of one variable in a TSD file.}
\usage{
read.TSD_readAllTimeStepsOfOneVariable(numtInFile, lvar, var, con, index, dtyp,
  valid_datatypes, dtyp_present, datasizes, endian, arecomplex)
}
\value{

}
\description{
(Internal) Read all time steps of one variable in a TSD file.
}
\examples{
\dontrun{}

}
