% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.na.R
\name{rm.na}
\alias{rm.na}
\title{Removing NAs, returning a vector. Not used since na.omit() serves the same purpose.}
\usage{
rm.na(x)
}
\arguments{
\item{x}{Input object.}
}
\description{
Removing NAs, returning a vector. Not used since na.omit() serves the same purpose.
}
\examples{
\dontrun{}

}
