% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dhms.R
\name{s2dhms}
\alias{s2dhms}
\title{Transforms seconds to an array of days, hours, minutes and seconds.}
\usage{
s2dhms(sec, digits = 3, clock.out = FALSE, names = c("d", "h", "m", "s"),
  strip = FALSE, drop = FALSE)
}
\arguments{
\item{sec}{is the input seconds.}

\item{digits}{is the number of digits of the "seconds" part of the output.}

\item{clock.out}{is TRUE if the time excluding the days are to be returned as a string vector of elements HH:MM:SS.FFF (two digits).}
}
\value{

}
\description{
Transforms seconds to an array of days, hours, minutes and seconds.
}
\examples{
\dontrun{}

}
