% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitSeqIntoBlocks.R
\name{splitSeqIntoBlocks}
\alias{splitSeqIntoBlocks}
\title{Splits a vector into groups based on the size of one element and a size limit.}
\usage{
splitSeqIntoBlocks(t, size1, size, blocks = 1)
}
\arguments{
\item{t}{is a vector}

\item{size1}{the size of each element.}

\item{size}{the maximum size each block.}

\item{blocks}{(optional) if the number of blocks is lower than 'blocks', distribute evenly.}
}
\value{

}
\description{
Splits a vector into groups based on the size of one element and a size limit.
}
\examples{
\dontrun{}

}
