% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strff.R
\name{strff}
\alias{strff}
\title{Returns TRUE if any of the strings of 'first' occur as the first characters of the strings of 'str'.}
\usage{
strff(first, str, n = NULL, ignore.case = TRUE)
}
\arguments{
\item{first}{is the string holding the 'n' characters to find as starting characters in 'str'. Only the first element of 'first' is used.}

\item{str}{is string in which to find 'first' as starting characters.}

\item{n}{is the number of leading characters, defaulted to the length of 'first'.}

\item{ignore.case}{is FALSE for case sensitivity.}
}
\value{

}
\description{
Returns TRUE if any of the strings of 'first' occur as the first characters of the strings of 'str'.
}
\examples{
\dontrun{}

}
