% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateHeaderTSD.R
\name{UpdateHeaderTSD}
\alias{UpdateHeaderTSD}
\title{Updates headers of TSD files for which extra rows of zeros was reserved using the option 'reserve' in write.TSD. If fewer rows of zeros was reserved, than the number of time steps to append to the TSD file, UpdateHeaderTSD() will cause trouble.}
\usage{
UpdateHeaderTSD(filename, lvar, pos, endian = "big")
}
\arguments{
\item{filename}{is the name of the TSD file for which the header is to be updated.}

\item{lvar}{is the variable lengths vector given as c(lvar_1_1, lvar_1_2, ... , lvar_1_nvar, lvar_2_1, lvar_2_2, ... , lvar_2_nvar, ... , lvar_numt_1, lvar_numt_2, ... , lvar_numt_nvar).}

\item{pos}{is the position in bytes from the start of the file, at which the header is to be updated.}

\item{endian}{is the endian of the existing TSD file.}
}
\description{
Updates headers of TSD files for which extra rows of zeros was reserved using the option 'reserve' in write.TSD. If fewer rows of zeros was reserved, than the number of time steps to append to the TSD file, UpdateHeaderTSD() will cause trouble.
}
\examples{
\dontrun{}

}
