% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utim.TSD.R
\name{utim.TSD}
\alias{utim.TSD}
\alias{ftim.TSD}
\alias{mtim.TSD}
\alias{utim2mtim}
\alias{utim2ftim}
\alias{utim2FILETIME}
\alias{ftim2list}
\alias{ftim2mtim}
\alias{ftim2utim}
\alias{mtim2FILETIME}
\alias{mtim2ftim}
\alias{mtim2utim}
\alias{FILETIME2mtim}
\alias{FILETIME2utim}
\alias{interpret.mtim}
\title{Transforms time input in TSD format to unix time. If Matlab time 'mtim' is given, it is assumed to be in UTC.}
\usage{
utim.TSD(data, keep.list = FALSE, ...)

ftim.TSD(data, keep.list = FALSE, ...)

mtim.TSD(data, keep.list = FALSE, ...)

utim2mtim(x, ...)

utim2ftim(x, format = "yyyymmddHHMMSS.FFF", digits = 3, indt = NULL,
  split = FALSE, numeric = FALSE, ...)

utim2FILETIME(x, xBase = -11644473600, tz = "UTC")

ftim2list(x, ...)

ftim2mtim(x, ...)

ftim2utim(x, ...)

mtim2FILETIME(x, xBase = -11644473600, tz = "UTC")

mtim2ftim(x, format = "yyyymmddHHMMSS.FFF", digits = 3, indt = NULL,
  split = FALSE, numeric = FALSE, ...)

mtim2utim(x, ...)

FILETIME2mtim(x, xBase = -11644473600, tz = "UTC")

FILETIME2utim(x, xBase = -11644473600, tz = "UTC")

interpret.mtim(x)
}
\arguments{
\item{data}{is a list with names as used in the TSD file format (usually from read.TSD()).}

\item{keep.list}{is TRUE if 'ftim' is to be returned as a list even though only one vector of 'ftim' is extracted.}

\item{...}{arguments passed from or to other functoins.}

\item{x}{are the input time points. For conversions from ftim to other units, POSIXct is supported (such as the value returned from Sys.time()).}

\item{format}{is a string specifying the format of the return values. "yyyy" denotes year, "mm" denotes month, "dd" denotes day, "HH" denotes hour, "MM" denotes minutes, "SS" denotes seconds and "FFF" denotes fractional seconds of 'digits' digits. (Similar to the MATLAB function timestr()). Also an index number given by 'indt' can be added in the output, as well as text and separators, such as "yyyy-mm-dd\\nHH:MM:SS.FFF\\nPing: indt", which separates the date, time, and ping number in acoustic data by line spaces.}

\item{digits}{is the number of digits to return after seconds.}

\item{indt}{is the optional index number to add to the output.}

\item{split}{is TRUE if the formated time points are to be returned as a matrix of columns "D" and "T".}

\item{numeric}{is TRUE if the time values are to be returned as a numeric column matrix (only in the case split==TRUE).}

\item{xBase}{is the base of Windows FILETIME: xBase=unclass(as.POSIXct('1601-1-1', tz="UTC"))[1].}

\item{tz}{is the the time zone. Overrides 'xBase' if not given as tz="UTC". See as.POSIXlt().}
}
\value{

}
\description{
\code{utim.TSD} Transforms time input in TSD format to unix time. If Matlab time 'mtim' is given, it is assumed to be in UTC.
\code{ftim.TSD} Transforms time input in TSD format to formatted time D=yyyymmdd and T=HHMMSS.FFF in a column matrix or vector.
\code{mtim.TSD} Transforms time input in TSD format to UTC Matlab time.
\code{utim2mtim} Transforms unix time to Matlab time, which will be in UTC.
\code{utim2ftim} Transforms unix time to formatted time D=yyyymmdd and T=HHMMSS.FFF in a column matrix or vector.
\code{utim2FILETIME} Transforms UNIX time to Windows FILETIME (the number of 100 nanoseconds since 1 January 1601).
\code{ftim2list} Transforms 'ftim' in the TSD file format to a list used by ftim2mtim() and ftim2utim().
\code{ftim2mtim} Transforms 'ftim' in the TSD file format to UTC Matlab time.
\code{ftim2utim} Transforms 'ftim' in the TSD file format to unix time.
\code{mtim2FILETIME} Transforms MATLAB serial date (Number of days since Jesus was born) to Windows FILETIME (the number of 100 nanoseconds since 1 January 1601).
\code{mtim2ftim} Transforms unix time to formatted time D=yyyymmdd and T=HHMMSS.FFF in a column matrix or vector.
\code{mtim2utim} Transforms Matlab time in UTC to unix time.
\code{FILETIME2mtim} Transforms Windows FILETIME to UTC Matlab time.
\code{FILETIME2utim} Transforms Windows FILETIME to UNIX time.
\code{interpret.mtim} Transforms time given either as MATLAB serial date number 'mtim' (number of days since January 0, 0000), 'utim' (number of seconds elapsed since UTC 00:00, 1970-01-01), or 'ftim' (yyyyddHHMMSS.FFF or yyyymmddSSSSS.FFF) to MATLAB serial date number.
}
\examples{
\dontrun{}

}
