% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecRealImaginary2complex.R
\name{vecRealImaginary2complex}
\alias{vecRealImaginary2complex}
\title{Converts to complex from data stored with the real in the first half and the imaginary in the second part of the vector 'x'.}
\usage{
vecRealImaginary2complex(x)
}
\arguments{
\item{x}{is a vector where the first half is the real part and the second half is the imaginary part.}
}
\description{
Converts to complex from data stored with the real in the first half and the imaginary in the second part of the vector 'x'.
}
\examples{
\dontrun{}

}
