% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WinPath.R
\name{WinPath}
\alias{WinPath}
\title{Run WinPath(), paste in a Windows file path, hit Enter, and get a valid path out (changing \ to /).}
\usage{
WinPath()
}
\value{

}
\description{
Run WinPath(), paste in a Windows file path, hit Enter, and get a valid path out (changing \ to /).
}
