% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.TSD_collapse_string.R
\name{write.TSD_collapse_string}
\alias{write.TSD_collapse_string}
\title{Function used for collapsing string vectors, arrays or lists to strings of length 1 at each time step (only used on elements of the data that are recognized as string type).}
\usage{
write.TSD_collapse_string(y, numt, split_char = "\\037", ts = "last")
}
\arguments{
\item{y}{is a string vector to collapse into a single string with split_char filled in between the vector elements.}

\item{numt}{is the number of time steps.}

\item{split_char}{is a character used to split strings into vector elemens when restoring the vector.}

\item{ts}{is a numeric vector specifying the rules for regarding the last dimensions of arrays as time: If the number of dimensions of an array is included in the set 'ts' the last dimension of that array is considered to be along time, so that if ts=3 and we wish to write an array of dimension [3, 4, 2], this array will be written as two [3, 4] arrays. 'ts' is utilized through seq_along(dim(x))[ts], so that if ts=-2 matrices are not regarded as having time along the second dimension.}
}
\description{
Function used for collapsing string vectors, arrays or lists to strings of length 1 at each time step (only used on elements of the data that are recognized as string type).
}
\examples{
\dontrun{}

}
