% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.TSD_d000_lt2v.R
\name{write.TSD_d000_lt2v}
\alias{write.TSD_d000_lt2v}
\title{List representation of dimension data as a function of time steps -> vector representation.}
\usage{
write.TSD_d000_lt2v(d000l, numt)
}
\arguments{
\item{d000l}{is a list of dimension information, as returnerd from write.TSD_get_d000_lv().}

\item{numt}{is the number of time steps.}
}
\value{

}
\description{
List representation of dimension data as a function of time steps -> vector representation.
}
\examples{
\dontrun{}

}
