% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.TSD_d000_lv2lt.R
\name{write.TSD_d000_lv2lt}
\alias{write.TSD_d000_lv2lt}
\title{List representation as a function of variables -> list representation as a function of time steps.}
\usage{
write.TSD_d000_lv2lt(d000l, numt)
}
\arguments{
\item{d000l}{is a list of dimension information, as returnerd from write.TSD_get_d000_lv().}

\item{numt}{is the number of time steps.}
}
\value{

}
\description{
List representation as a function of variables -> list representation as a function of time steps.
}
\examples{
\dontrun{}

}
