% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.TSD_fun_d000.R
\name{write.TSD_fun_d000}
\alias{write.TSD_fun_d000}
\title{Function extracting the dimensions of the variable (possibly list) 'y'.}
\usage{
write.TSD_fun_d000(y, numt, ts, ind.out = FALSE)
}
\arguments{
\item{y}{is an R object accepted by write.TSD().}

\item{numt}{is the number of time steps.}

\item{ts}{is a numeric vector specifying the rules for regarding the last dimensions of arrays as time: If the number of dimensions of an array is included in the set 'ts' the last dimension of that array is considered to be along time, so that if ts=3 and we wish to write an array of dimension [3, 4, 2], this array will be written as two [3, 4] arrays. 'ts' is utilized through seq_along(dim(x))[ts], so that if ts=-2 matrices are not regarded as having time along the second dimension.}

\item{ind.out}{Logical: If TRUE output indices instead of the dimensions.}
}
\description{
Function extracting the dimensions of the variable (possibly list) 'y'.
}
\examples{
\dontrun{}

}
