% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.TSD_fun_numt.R
\name{write.TSD_fun_numt}
\alias{write.TSD_fun_numt}
\title{Function to be used for extracting the number of time steps 'numt' if not given. The function 'fun_numt' is applied to each list element of 'x'. For list variables of 'x' the length is extracted, for matrices the number of columns is extracted and for vectors 'numt' is set to 1.}
\usage{
write.TSD_fun_numt(y, ts)
}
\arguments{
\item{y}{is an R object accepted by write.TSD().}

\item{ts}{is a numeric vector specifying the rules for regarding the last dimensions of arrays as time: If the number of dimensions of an array is included in the set 'ts' the last dimension of that array is considered to be along time, so that if ts=3 and we wish to write an array of dimension [3, 4, 2], this array will be written as two [3, 4] arrays. 'ts' is utilized through seq_along(dim(x))[ts], so that if ts=-2 matrices are not regarded as having time along the second dimension.}
}
\description{
Function to be used for extracting the number of time steps 'numt' if not given. The function 'fun_numt' is applied to each list element of 'x'. For list variables of 'x' the length is extracted, for matrices the number of columns is extracted and for vectors 'numt' is set to 1.
}
\examples{
\dontrun{}

}
