% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.TSD_get_ts.R
\name{write.TSD_get_ts}
\alias{write.TSD_get_ts}
\title{(Internal:) Get the 'ts' object, which defines along which dimension time steps are organized.}
\usage{
write.TSD_get_ts(d, ts, numt = NULL, ischar = FALSE)
}
\arguments{
\item{d}{is a vector of the dimensions of the array.}

\item{ts}{(see write.TSD).}

\item{numt}{is the number of time steps.}

\item{ischar}{is TRUE if the variable is character, in which case all string elements are collapsed to one vector if 'numt' is given in write.TSD() and does not fit the length of the string vector.}
}
\value{

}
\description{
(Internal:) Get the 'ts' object, which defines along which dimension time steps are organized.
}
\examples{
\dontrun{}

}
