% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.TSD_lvar_d000.R
\name{write.TSD_lvar_d000}
\alias{write.TSD_lvar_d000}
\title{Simple function for extracting the lengths of the dimension variable at each time step.}
\usage{
write.TSD_lvar_d000(d000, numt)
}
\arguments{
\item{d000}{is a vector of dimension information stored as c(narrays_1, ndims_1, indx_1, dims_1, narrays_2, ndims_2, indx_2, dims_2, ... , narrays_numt, ndims_numt, indx_numt, dims_numt), where 'narrays_p' is the number of variables at time step 'p' that have dimension (vectors are not regarded av having dimension, corresponding to the dim() function), 'ndims_p' is the number of dimensions for each variable at time step 'p', 'indx_p' is the indices for the variables that posess dimension at time step 'p', and 'dims_p' are the dimension values for the variables indexed by 'indx_p' at time step 'p'..}

\item{numt}{is the number of time steps.}
}
\description{
Simple function for extracting the lengths of the dimension variable at each time step.
}
\examples{
\dontrun{}

}
