% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeropad.R
\name{zeropad}
\alias{zeropad}
\title{Pads numerics with zeros at the beginning.}
\usage{
zeropad(x, n = max(nchar(x)))
}
\arguments{
\item{x}{is a numeric vector.}

\item{n}{is the number of characters in the returned strings.}
}
\description{
Pads numerics with zeros at the beginning.
}
\examples{
\dontrun{}

}
