% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros.R
\name{chars}
\alias{chars}
\alias{FALSEs}
\alias{NaNs}
\alias{NAs}
\alias{ones}
\alias{zeros}
\title{Generate arrays with the given dimenstions.}
\usage{
chars(...)

FALSEs(...)

NaNs(...)

NAs(...)

ones(..., type = "double")

zeros(..., type = "double")
}
\arguments{
\item{...}{the elements specifying the dimension of the array.}

\item{type}{specifies the return data type as either "int" or "double" (default), used in zeros() and ones().}
}
\value{
An array of zeros, ones, NAs, NaNs, empry characters or FALSEs, with the given dimension.
}
\description{
Generate arrays with the given dimenstions.
}
\examples{
\dontrun{
zeros(2:3)
ones(2:3, 4)
chars(2:3)
FALSEs(2:3)
NAs(2:3)
NaNs(2:3)}

}
