% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labl.TSD.R
\name{labl.TSD}
\alias{labl.TSD}
\title{Returns the variable names required in echoIBM() and related functions, given as four character strings of names as stored in the TSD format. The aim of this function is to avoid errors due to misspelling.}
\usage{
labl.TSD(var = NULL, adds = NULL, list.out = TRUE)
}
\arguments{
\item{var}{The variable classes for which variable names are requested. Use names(labl.TSD()) to see implemented variable classes are.}

\item{adds}{A vector of additional variables to include.}

\item{list.out}{Logical: If FALSE, unlist the output when multiple elements are requested in var.}
}
\value{

}
\description{
Returns the variable names required in echoIBM() and related functions, given as four character strings of names as stored in the TSD format. The aim of this function is to avoid errors due to misspelling.
}
\examples{
\dontrun{}

}
