% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papply.R
\name{papply}
\alias{papply}
\title{Parallel version of lapply}
\usage{
papply(X, FUN, ..., cores = 1, outfile = "", text = "Processing... ",
  msg = TRUE)
}
\arguments{
\item{X, FUN, ...}{See lapply.}

\item{cores}{An integer giving the number of cores to run the function FUN over.}

\item{outfile}{Set this to FALSE to suppress printing from the cores.}
}
\value{

}
\description{
Detects and opens cores for simpler parallel lapply.
}
\examples{
f <- function(i){
	for(j in 1:3){
		for(k in 1:500){
			Sys.sleep(0.001)
			d <- runif(100)
		}
	}	
}
system.time(p <- papply(1:4, f, cores=1))
system.time(p <- papply(1:4, f, cores=2))

}
