% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyIntegers.R
\name{prettyIntegers}
\alias{prettyIntegers}
\title{Compressed formating of an integer vector.}
\usage{
prettyIntegers(x, force = FALSE, collapse = ", ", sep = "...")
}
\arguments{
\item{x}{An integer vector.}

\item{force}{Logical: If TRUE the new string should be returned regardless of whether it has more characters than the original.}

\item{collapse}{The string to separate numbers by.}

\item{sep}{The string to separate numbers by when printing ranges.}
}
\value{

}
\description{
Compressed formating of an integer vector.
}
\examples{
\dontrun{}

}
