% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.TSD_funstring2fun.R
\name{read.TSD_funstring2fun}
\alias{read.TSD_funstring2fun}
\title{Simple function for extracting the function saved as a string using \code{\link{function2character}}).}
\usage{
read.TSD_funstring2fun(x)
}
\arguments{
\item{x}{is a function string.}
}
\description{
Simple function for extracting the function saved as a string using \code{\link{function2character}}).
}
\examples{
\dontrun{
f <- function(s,r) s+2*r
f <- function2character(f, "FUNCTION")
read.TSD_funstring2fun(f)(13, 4)
}

}
