% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.TSD_getindtFirstLast.R
\name{read.TSD_getindtFirstLast}
\alias{read.TSD_getindtFirstLast}
\title{A function for extracting the first and last value of 'indt'.}
\usage{
read.TSD_getindtFirstLast(labl, con, index, datatypes, datasizes,
  endian = .Platform$endian)
}
\arguments{
\item{labl}{is a vector of the TSD labels of the file.}

\item{con}{is a connection to a binary file, as returned by file(x,"rb").}

\item{index}{is a matrix of the positions in the type of the variables and time steps.}

\item{datatypes}{The data types of each element of 'labl'.}

\item{datasizes}{The data sizes of each element of 'labl'.}

\item{endian}{is the endianness of the file, changed if the float is not read properly.}
}
\value{

}
\description{
A function for extracting the first and last value of 'indt'.
}
\examples{
\dontrun{}

}
